refactor(examples): split examples between libraries
