+++ examples/salvo-echo/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "salvo-echo"
+version = "0.6.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+engineioxide = { path = "../../engineioxide", feature = ["hyper-v1"] }
+salvo = { version = "0.58.2", features = ["tower-compat"] }
+tokio = { version = "1.33.0", features = ["full"] }
+tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
+tracing = "0.1.37"
+serde = "1.0.188"
+serde_json = "1.0.107"
+futures = "0.3.28"
+
+[[bin]]
+name = "salvo-echo"
+path = "salvo_echo.rs"
