+++ examples/warp-echo/Cargo.toml
@@ -6,27 +6,16 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-engineioxide = { path = "../../engineioxide" }
-axum = { version = "0.6.20" }
+socketioxide = { path = "../../socketioxide" }
 warp = { version = "0.3.6" }
 hyper = { version = "0.14.27" }
 tokio = { version = "1.33.0", features = ["full"] }
-tower = { version = "0.4.13" }
-tower-http = { version = "0.4.4", features = ["cors"] }
 tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
 tracing = "0.1.37"
 serde = "1.0.188"
 serde_json = "1.0.107"
 futures = "0.3.28"
 
-[[example]]
-name = "engineio-axum-echo"
-path = "src/axum_echo.rs"
-
-[[example]]
-name = "engineio-hyper-echo"
-path = "src/hyper_echo.rs"
-
-[[example]]
-name = "engineio-warp-echo"
-path = "src/warp_echo.rs"
+[[bin]]
+name = "warp-echo"
+path = "warp_echo.rs"
